Ceed

LibCEED.Ceed — Type

Ceed(spec="/cpu/self")

Wraps a libCEED Ceed object, created with the given resource specification string.

LibCEED.getresource — Function

getresource(c::Ceed)

Returns the resource string associated with the given Ceed object.

LibCEED.get_preferred_memtype — Function

get_preferred_memtype(c::Ceed)

Returns the preferred MemType (either MEM_HOST or MEM_DEVICE) of the given Ceed object.

LibCEED.isdeterministic — Function

isdeterministic(c::Ceed)

Returns true if backend of the given Ceed object is deterministic, and false otherwise.

LibCEED.iscuda — Function

iscuda(c::Ceed)

Returns true if the given Ceed object has resource "/gpu/cuda/*" and false otherwise.