Constants and Enumerations

LibCEED.CeedScalar — Type

CeedScalar

Scalar (floating point) type. Typically equivalent to Float64, but libCEED can be configured at compile-time to use Float32. See also get_scalar_type.

LibCEED.CeedInt — Type

CeedInt

Integer type, used for indexing. Equivalent to Int32.

LibCEED.CeedInt8 — Type

CeedInt8

Integer type, for small integers. Equivalent to Int8.

LibCEED.CeedSize — Type

CeedSize

Integer type, used for array sizes. Equivalent to Int.

LibCEED.QuadMode — Type

QuadMode

One of GAUSS or GAUSS_LOBATTO.

LibCEED.MemType — Type

MemType

One of MEM_HOST or MEM_DEVICE.

LibCEED.CopyMode — Type

CopyMode

One of COPY_VALUES, USE_POINTER or OWN_POINTER.

OWN_POINTER is not typically supported for objects created in Julia, because those must be destroyed by the garbage collector, and cannot be freed from C.

LibCEED.EvalMode — Type

EvalMode

Evaluation mode used in the specification of input and output fields for Q-functions, e.g. in @interior_qf.

One of:

EVAL_NONE

EVAL_INTERP

EVAL_GRAD

EVAL_DIV

EVAL_CURL

EVAL_WEIGHT

See the libCEED documentation for further information.

LibCEED.TransposeMode — Type

TransposeMose

Denotes whether a linear transformation or its transpose should be applied. Either NOTRANSPOSE or TRANSPOSE.

LibCEED.NormType — Type

NormType

Denotes type of vector norm. One of NORM_1, NORM_2, or NORM_MAX.

LibCEED.Topology — Type

Topology

Type of basis shape to create non-tensor H1 element basis. One of LINE, TRIANGLE, QUAD, TET, PYRAMID, PRISM, or HEX.

The dimension can be extracted with bitshift:

dim = Int(topology) >> 16

LibCEED.STRIDES_BACKEND — Constant

STRIDES_BACKEND

Indicate that the stride is determined by the backend.